fixed compilation issue (only on some machine for some reason) caused by a double linking of ImGui.
Editor compiles but is not functional yet.

+++ OsmiumCore/AssetManagement/AssetType/TextureAsset.cpp
@@ -24,7 +24,7 @@ std::mutex & TextureAsset::GetRessourceMutex() {
     return Resources::ResourceManager::getResourceMutex(Resources::ResourceType_Texture);
 }
 
-unsigned long TextureAsset::GetTextureHandle() const {
+uint32_t TextureAsset::GetTextureHandle() const {
     return textureHandle;
 }
